Some improvements
* Added changeRepeatMode to RemoteCommand
* Added toleranceBefore/After as configuration options for more accurate seeking in VBR mp3s
* Added seek(to:toleranceBefore:toleranceAfter)

+++ SwiftAudio/Classes/AVPlayerWrapper/AVPlayerWrapperDelegate.swift
@@ -13,7 +13,7 @@ protocol AVPlayerWrapperDelegate: class {
     func AVWrapper(didChangeState state: AVPlayerWrapperState)
     func AVWrapper(secondsElapsed seconds: Double)
     func AVWrapper(failedWithError error: Error?)
-    func AVWrapper(seekTo seconds: Int, didFinish: Bool)
+    func AVWrapper(seekTo seconds: Double, didFinish: Bool)
     func AVWrapper(didUpdateDuration duration: Double)
     func AVWrapperItemDidPlayToEndTime()
     func AVWrapperDidRecreateAVPlayer()
